infrastructure/requirements.txt
aws-cdk-lib>=2.240.0,<3.0.0
constructs>=10.5.0,<11.0.0
boto3>=1.28.0
